+++ src/main/java/net/unicon/cas/addons/info/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Various classes providing insights into CAS server runtime information
+ */
+package net.unicon.cas.addons.info;
+