+++ changelog.md
@@ -5,6 +5,7 @@
 =====================================
 
 * added [MongoServiceRegistryDao](https://github.com/Unicon/cas-addons/blob/master/src/main/java/net/unicon/cas/addons/serviceregistry/mongodb/MongoServiceRegistryDao.java)
+* added various classes providing CAS server runtime information in the [info](https://github.com/Unicon/cas-addons/blob/master/src/main/java/net/unicon/cas/addons/info) package
 
 ###Changes in version 1.0 (2012-10-07)
 =====================================